Important modifications to be applied to Workshop 6
Finished Issue #7

Fixed the few slides (8;49;58;89).

Tried to reduce long sentences or long slides.

Added a few presenter notes and made sure they were prensent in both EN and FR

For the Hosmer-Lemenshow test slide: it was already excluded from both presentations, so I only added a presenter note on this test so the presenter can refer to it if needed.